Moving types under printer package
