+++ pkg/printer/cluster.go
@@ -2,7 +2,7 @@ package printer
 
 import (
 	"fmt"
-	"github.com/cnoe-io/idpbuilder/pkg/types"
+	"github.com/cnoe-io/idpbuilder/pkg/printer/types"
 	"io"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
